Release 1.0
With more_up_packs you're now able to have more health and breath points in minetest, as long as both the upgrade_packs mod by Krock and the mtg_plus mod by Wuzzy are both installed and active. The former is the base framework that allows this mod to function in the first place. The latter is used for the final +50HP upgrade.
Upgrades range from +20HP to +50HP in steps of 10. There's currently also one upgrade that allows you to have +10 Breath so you can survive even longer unter water.